+++ src/main/cxx/core/optimizer.h
@@ -898,7 +898,7 @@ namespace especia {
         /**
          * The random number generator.
          */
-        const Normal_Deviate<Mt19937> deviate;
+        const Normal_Deviate<Mt19937_32> deviate;
     };
 
 }