+++ NEWS.md
@@ -4,8 +4,10 @@ The main programs were renamed and a few utility programs were added.
 The source code, documentation, and ancillary files were revised.
 In addition, Version 2020.1 provides several new features and enhancements:
 
-* Especia now uses 64-bit instead of 32-bit random seeds.
-* PCG is included in addition to Mersenne Twister. 
+* Especia now uses 64-bit instead of 32-bit seeds to initialialize the Mersenne
+Twister random number generator.
+* A PCG random number generator is included with the source code, in addition
+to the Mersenne Twister.
 * Especia uses all principal axes of the mutation ellipsoid to estimate parameter
 uncertainties.
 * Especia executes a super-sampled computation of the instrumental convolution